+++ seantis/people/upgrades.py
@@ -1,0 +1,13 @@
+from plone import api
+
+
+def run_import_step_from_profile(
+    context, step, module='seantis.people', profile='default'
+):
+    setup = api.portal.get_tool('portal_setup')
+    profile = 'profile-{}:{}'.format(module, profile)
+    setup.runImportStepFromProfile(profile, step)
+
+
+def install_custom_controlpanel(context):
+    run_import_step_from_profile(context, 'controlpanel')